notify: add email integration tests
